% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateHarmoStatus}
\alias{annotateHarmoStatus}
\title{Set variable annotation with Harmonization Status term}
\usage{
annotateHarmoStatus(tibble, variables = NULL, status)
}
\arguments{
\item{tibble}{Tibble to be annotated.}

\item{variables}{A character vector of variable names to be annotated. If NULL or empty, all
the columns of the tibble will be annotated.}

\item{status}{The harmonization status to apply: 'complete', 'undetermined' or 'impossible'. If NULL, the annotation will be deleted.}
}
\value{
The annotated tibble
}
\description{
Apply or remove an harmonization status annotation from a set of variables. The harmonization status
is described by the "status" vocabulary in the "Mlstr_harmo" taxonomy.
}
