% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHarmonizationBook.R
\name{.getTemplate}
\alias{.getTemplate}
\title{Get item (vocabulary, term, variable, categories, category) template}
\usage{
.getTemplate(type, dir)
}
\description{
Look for R markdown template file from provided directory (from _templates sub-directory) or from
the package default ones.
}
\keyword{internal}
