% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOpalTable.R
\name{getOpalTable}
\alias{getOpalTable}
\title{Get a Opal table as a tibble}
\usage{
getOpalTable(opal, project, table, variables = NULL)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table is located.}

\item{table}{Table name from which the tibble should be extracted.}

\item{variables}{List of variable names or Javascript expression that selects the variables of a table (ignored if value does not refere to a table). See javascript documentation: http://wiki.obiba.org/display/OPALDOC/Variable+Methods}
}
\description{
Shortcut function to assign a Opal table to a tibble in the R server-side session
and then retrieve it into the R client-side session. Requires to have the permission to
see the individual values of the table and to perform R assignments.
}
