% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHarmonizationBook.R
\name{makeHarmonizationBook}
\alias{makeHarmonizationBook}
\title{Make a R markdown book}
\usage{
makeHarmonizationBook(
  opal,
  project,
  table,
  taxonomy = "Mlstr_area",
  vocabularies = NULL,
  locale = "en",
  outDir = file.path(getwd(), "_harmobook"),
  outFiles = list(domains = "04-domains.Rmd", index = "06-index.Rmd")
)
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table is located.}

\item{table}{Table name from which the data dictionary should be extracted.}

\item{taxonomy}{Taxonomy name to group variables by vocabularies and terms.}

\item{vocabularies}{Vocabularies that shall be included, default is all (NULL value).}

\item{locale}{The locale in which the labels should be extracted (default is "en").}

\item{outDir}{Output folder where the R markdown files will be produced (default is
'_harmobook' folder in the current working directory). Also used to find R markdown templates (in the
'_templates' subfolder); if non is found the 'harmor' package default ones will be used.}

\item{outFiles}{Names of the output files, idenitified by "domains" or "index".}
}
\description{
Get the variables from a Opal table and dump a book of R markdown templates,
ready for the including derivation code chunks and harmonization status.
}
