% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOpalTable.R
\name{saveOpalTable}
\alias{saveOpalTable}
\title{Save the tibble as a Opal table}
\usage{
saveOpalTable(opal, tibble, project, table, overwrite = TRUE,
  force = FALSE, identifiers = NULL, policy = "required",
  id.name = "id", type = "Participant", variables = NULL,
  categories = NULL)
}
\arguments{
\item{opal}{Opal connection object.}

\item{tibble}{The tibble object to be imported.}

\item{project}{Project name where the table will be located.}

\item{table}{Destination table name.}

\item{overwrite}{If the destination table already exists, it will be replaced (deleted and then
imported). Otherwise the table will be updated (data dictionaries merge may conflict). Default is TRUE.}

\item{force}{If the destination already exists, stop with an informative message if this flag is FALSE (default).}

\item{identifiers}{Name of the identifiers mapping to use when assigning entities to Opal.}

\item{policy}{Identifiers policy: 'required' (each identifiers must be mapped prior importation (default)), 'ignore' (ignore unknown identifiers) and 'generate' (generate a system identifier for each unknown identifier).}

\item{id.name}{The name of the column representing the entity identifiers. Default is 'id'.}

\item{type}{Entity type (what the data are about). Default is 'Participant'}

\item{variables}{A data frame with one row per variable (column name) and then one column per property/attribute (Opal Excel format).}

\item{categories}{A data frame with one row per category (columns variable and name) and then column per property/attribute (Opal Excel format). If there are
no categories, this parameter is optional.}
}
\value{
An invisible logical indicating whether the destination table exists.
}
\description{
Assign the tibble from the R client-side session to the R session-side session with the provided
symbol name and import it into the Opal project. If variables dictionary is provided, it will be applied after
the import.
}
