% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHarmonizationBook.R
\name{makeHarmonizationBook}
\alias{makeHarmonizationBook}
\title{Make a R markdown book}
\usage{
makeHarmonizationBook(opal, project, table, taxonomy = "Mlstr_area",
  locale = "en", outDir = file.path(getwd(), "_harmobook"),
  outFile = "04-domains.Rmd")
}
\arguments{
\item{opal}{Opal connection object.}

\item{project}{Project name where the table is located.}

\item{table}{Table name from which the data dictionary should be extracted.}

\item{taxonomy}{Taxonomy name to group variables by vocabularies and terms.}

\item{locale}{The locale in which the labels should be extracted (default is "en").}

\item{outDir}{Output folder where the R markdown templates should be produced (default is
'_harmobook' folder in the current working directory).}

\item{outFile}{Name of the output file.}
}
\description{
Get the variables from a Opal table and dump a book of R markdown templates,
ready for the including derivation code chunks and harmonization status.
}
